+++ test/phoenix/tracker/shard_replication_test.exs
@@ -5,6 +5,7 @@ defmodule Phoenix.Tracker.ShardReplicationTest do
   @primary :"primary@127.0.0.1"
   @node1 :"node1@127.0.0.1"
   @node2 :"node2@127.0.0.1"
+  @moduletag :capture_log
 
   setup config do
     tracker = config.test
@@ -44,7 +45,7 @@ defmodule Phoenix.Tracker.ShardReplicationTest do
     resume_gossips(shard)
     # primary sends transfer_req to node1 after seeing behind
     ref = assert_transfer_req to: @node1, from: @primary
-    # node1 fulfills tranfer request and sends transfer_ack to primary
+    # node1 fulfills transfer request and sends transfer_ack to primary
     assert_transfer_ack ref, from: @node1
     assert_heartbeat to: @node1, from: @primary
     assert [{"node1", _}] = list(shard, topic)
@@ -387,6 +388,20 @@ defmodule Phoenix.Tracker.ShardReplicationTest do
 
     # Wait until primary is permanently down
     assert_receive {{:replica_permdown, @node1}, @primary}, permdown_period * 2
+  end
+
+  test "handle_info callback with bad return", %{shard_pid: shard_pid} do
+    Process.unlink(shard_pid)
+    ref = Process.monitor(shard_pid)
+    send(shard_pid, {:run, fn _state -> :bad end})
+    assert_receive {:DOWN, ^ref, :process, ^shard_pid, _}
+  end
+
+  test "handle_info callback with valid return", %{shard_pid: shard_pid} do
+    parent = self()
+    ref = make_ref()
+    send(shard_pid, {:run, fn _state -> {:noreply, send(parent, ref)} end})
+    assert_receive ^ref
   end
 
   ## Helpers